requirements.txt
fastapi
uvicorn
pydantic
openai
chromadb
python-dotenv
python-multipart
requests
